refactor(test dir): unify tests hierarchy
